pause file for popup
